Update SPDIF clock only when necessary
Previously the code would update the SPDIF clock setting at the
beginning of every command sent to the SD card. This included the math
needed to calculate the clock rates and clock dividers.

This commit makes it so that these calculations are only performed on
the first command and when the g_card_info->card_info.speed field has
been changed.

Can be disabled by setting SDIF_CLOCK_OPTIMIZE to 0.

Performance notes:
* This change produces bigger gains when the buffer size is small since
  those timings are dominated by the overhead of the serial SD commands
  and responses.
* I saw improvements for 1k buffer reads/writes of around 8%.