SD driver aggressively rounds CLKDIV up
This meant that trying to set the SD clock rate to 50MHz resulted in a
clock rate of 204/6 = 34MHz instead of the desired 50MHz. The way I
modified the code will result in SD clock rates which are a little
high:
  51MHz for high speed
  25.5MHz for the default speed
Hopefully that will be ok for most cards.

I hacked in a way to perform a 4MB read and it obtained a rate of
24.5MB/sec which was very close to the maximum of 25MB/sec for high
speed SD. It got my 16K reads up to 15.5MB/sec which isn't too bad
either.