Fail SD read/write if malloc() fails
These allocations are only performed if an unaligned (4-byte) buffer
makes it into disk_read() or disk_write().

Currently _sbrk() failures to find more heap for malloc() end up
trapping into the debugger anyway. So before a NULL could ever be
returned from malloc(), the code will be halted in the debugger. I
still added these checks incase this code is used in some other project
where this isn't the case or if Smoothie changes it malloc() failure
handling in the future.

Testing:
I forced malloc() to fail in the debugger and stepped through the code
during an unaligned read to make sure that it behaved as expected and
then copied the same code into the write path.